Move Device to its own action
