+++ src/libs/actions/App.js
@@ -16,7 +16,6 @@ import ROUTES from '../../ROUTES';
 import * as SessionUtils from '../SessionUtils';
 import getCurrentUrl from '../Navigation/currentUrl';
 import * as Session from './Session';
-import getDeviceID from '../getDeviceID';
 
 let currentUserAccountID;
 let currentUserEmail = '';
@@ -271,27 +270,6 @@ function openProfile() {
     Navigation.navigate(ROUTES.SETTINGS_PROFILE);
 }
 
-/**
- * Saves a unique deviceID into Onyx.
- */
-function setDeviceID() {
-    const connectionID = Onyx.connect({
-        key: ONYXKEYS.DEVICE_ID,
-        callback: (deviceID) => {
-            Onyx.disconnect(connectionID);
-            if (deviceID) {
-                Log.info('Found existing deviceID', false, deviceID);
-                return;
-            }
-
-            getDeviceID().then((uniqueID) => {
-                Log.info('Setting new deviceID', false, uniqueID);
-                Onyx.set(ONYXKEYS.DEVICE_ID, uniqueID);
-            });
-        },
-    });
-}
-
 export {
     setLocale,
     setSidebarLoaded,
@@ -299,5 +277,4 @@ export {
     openProfile,
     openApp,
     reconnectApp,
-    setDeviceID,
 };
